Random commit.  Making programs compile cleanly

+++ fprints.c
@@ -94,7 +94,7 @@ int main(int argc,char* argv[]) {
                 bzero(buf,sizeof(buf));
                 read(f,buf,SIGNATSIZE);
 
-                f=fnprint_compute(buf,(((unsigned int)buf) >> 24));
+                f=fnprint_compute(buf,((long int)buf >> 24));
 
                 if (f!=0xA120AD5C) { // Ignore only NOPs
                     printf("[%s+%d] %s ",argv[1],off,name);
